rewrite graph execution and preparation
Now resources are cached and kept alive until the reference size
changes.
The code around the graph has been refactored quite some, aiming
to reduce code complexity
