+++ nitrogen/Cargo.toml
@@ -8,7 +8,8 @@ edition = "2018"
 
 [features]
 default = ["winit_support"]
-winit_support = ["back/winit", "winit"]
+# winit_support = ["back/winit", "winit"]
+winit_support = ["gfx-backend-vulkan/winit", "winit"]
 x11 = []
 
 
@@ -22,19 +23,34 @@ bitflags = "1.0.4"
 smallvec = "0.6.5"
 winit = { version = "0.17.2", optional = true }
 
-[dependencies.gfx]
-package = "gfx-hal"
+#
+# [dependencies.gfx]
+# package = "gfx-hal"
+# git = "https://github.com/gfx-rs/gfx"
+# rev = "bd7f058efe78d7626a1cc6fbc0c1d3702fb6d2e7"
+# default-features = false
+
+# [dependencies.back]
+# package = "gfx-backend-vulkan"
+# git = "https://github.com/gfx-rs/gfx"
+# rev = "bd7f058efe78d7626a1cc6fbc0c1d3702fb6d2e7"
+# default-features = false
+
+# [dependencies.gfxm]
+# package = "gfx-memory"
+# git = "https://github.com/gfx-rs/gfx-memory"
+# rev = "e9bba2f935fe7b6e1fd2ed2862a0c5dcba8af364"
+
+[dependencies.gfx-hal]
 git = "https://github.com/gfx-rs/gfx"
 rev = "bd7f058efe78d7626a1cc6fbc0c1d3702fb6d2e7"
 default-features = false
 
-[dependencies.back]
-package = "gfx-backend-vulkan"
+[dependencies.gfx-backend-vulkan]
 git = "https://github.com/gfx-rs/gfx"
 rev = "bd7f058efe78d7626a1cc6fbc0c1d3702fb6d2e7"
 default-features = false
 
-[dependencies.gfxm]
-package = "gfx-memory"
+[dependencies.gfx-memory]
 git = "https://github.com/gfx-rs/gfx-memory"
 rev = "e9bba2f935fe7b6e1fd2ed2862a0c5dcba8af364"
